webaudio/the-audio-api/the-channelmergernode-interface/active-processing.https.html

<!doctype html>
<html>
  <head>
    <title>
      Test Active Processing for ChannelMergerNode
    </title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <script id="layout-test-code">
      // AudioProcessor that sends a message to its AudioWorkletNode whenver the
      // number of channels on its input changes.
      let filePath =
          '../the-audioworklet-interface/processors/active-processing.js';


      // AudioWorkletNode that verifies the number of channels matches
      // expectations.
      class ActiveProcessingTesterNode extends AudioWorkletNode {
        constructor(context, options) {
          super(context, 'active-processing-tester', options);
          this.port.onmessage = this.handleMessage_.bind(this);
          this.expectedValues_ = options.expectedValues;

          // The test harness function to use for testing.
          this.testHarness_ = options.testHarness;

          // Index into |expectedValue_| for comparing the actual and expected
          // number of channels.
          this.index_ = 0;
        }

        handleMessage_(event) {
          let count = event.data.channelCount;
          let finished = event.data.finished;

          // If we're finished, end testing and close context if we're not yet
          // closed.
          if (finished) {
            if (context.state != 'closed') {
              context.close();
            }

            // Verify that we got the expected number of changes.
            this.testHarness_.step(() => {
              assert_equals(
                  this.index_, this.expectedValues_.length,
                  'Number of distinct values');
            });
            this.testHarness_.done();
            return;
          }

          if (this.index_ >= this.expectedValues_.length) {
            this.testHarness_.unreached_func('Failed');
          } else {
            // Verify that the number of channels now matches the expected
            // number of channels.
            this.testHarness_.step(() => {
              assert_equals(
                  count, this.expectedValues_[this.index_],
                  `Test ${this.index_}: Number of convolver output channels:`);
            });
          }

          ++this.index_;
        }
      }

      let context;

      async function testActiveProcessing() {
        context = new AudioContext();

        await context.audioWorklet.addModule(filePath);

        const src = new OscillatorNode(context);

        // Number of inputs for the ChannelMergerNode.  Pretty arbitrary, but
        // should not be 1.
        const numberOfInputs = 7;
        const node =
            new ChannelMergerNode(context, {numberOfInputs: numberOfInputs});

        const testerNode = new ActiveProcessingTesterNode(context, {
          expectedValues: [numberOfInputs, 1],
          // Use as short a duration as possible to keep the test from taking
          // too much time.
          processorOptions: {testDuration: 1},
          testHarness: async_test(
              'ChannelMerger channel count changes to one when not actively processing'),
        });

        src.connect(node).connect(testerNode).connect(context.destination);
        src.start();
        // Stop the source after a short time so we can test the channel merger
        // change to not actively processing.
        src.stop(context.currentTime + .1);
      }

      testActiveProcessing();
    </script>
  </body>
</html>
